VectorSpin

Documentation for VectorSpin.

VectorSpin.BSplineAdvection

VectorSpin.Mesh

VectorSpin.H1fh!

VectorSpin.H2fh!

VectorSpin.H3fh!

VectorSpin.He!

VectorSpin.advection!

VectorSpin.bspline

VectorSpin.init

VectorSpin.integrate

VectorSpin.numeint

VectorSpin.step!

VectorSpin.step!

VectorSpin.step!

VectorSpin.step!

VectorSpin.step!

VectorSpin.step!

VectorSpin.translation!

VectorSpin.BSplineAdvection — Type

struct BSplineAdvection <: VectorSpin.AbstractAdvection

Advection to be computed on each row

mesh::Mesh

dims::Symbol

p::Int64

step::Float64

modes::Vector{Float64}

eig_bspl::Vector{Float64}

eigalpha::Vector{ComplexF64}

fhat::Matrix{ComplexF64}

VectorSpin.Mesh — Type

struct Mesh

Mesh type to store domain parameters

nv::Int64: Number of points in v

nx::Int64: Number of points in x

vmin::Float64: Domain size v ∈ ]vmin,vmax[

vmax::Float64: Domain size v ∈ ]vmin,vmax[

xmin::Float64: Domain size x ∈ [xmin,xmax]

xmax::Float64: Domain size x ∈ [xmin,xmax]

kx::Vector{Float64}: Wave number vector to compute derivative with FFTs

dx::Float64: Size step along x

dv::Float64: Size step along v

x::Vector{Float64}: points along x direction

v::Vector{Float64}: points along v direction

vnode::Vector{Float64}: center points along v direction

VectorSpin.H1fh! — Method

H1fh!(f0, f1, f2, f3, S1, S2, S3, dt, mesh, tiK)

compute the subsystem Hs1 M is even number

VectorSpin.H2fh! — Method

H2fh!(f0, f1, f2, f3, S1, S2, S3, t, mesh, tiK)

compute the subsystem Hs2

VectorSpin.H3fh! — Method

H3fh!(f0, f1, f2, f3, S1, S2, S3, t, mesh, tiK)

compute the subsystem Hs3

VectorSpin.He! — Method

He!(f0, f1, f2, f3, E1, t, H)

subsystem for He:

ft-Efv=0;

VectorSpin.advection! — Method

advection!(df, adv, v, dt)

interpolate df(x + delta) with Parabolic Spline Method (PSM)

We consider a linear advection problem in $p$ direction

\[\frac{\partial f}{\partial t} + a \frac{\partial f}{\partial x} =0.\]

From the conservation of the volume, we have the following identity

\[f_{j,\ell}(t)=\frac{1}{\Delta p} \int_{p_{\ell-1/2}} ^{p_{\ell+1/2}} f(x_j,p,t)\mathrm{d}{p} =\frac{1}{\Delta p} \int_{p_{\ell-1/2}-at} ^{p_{\ell+1/2}-at} f(x_j,p,0)\mathrm{d}{p}.\]

For simplicity, denote by $q\in [1, M]$ the index such that $p_{\ell+1/2}-at \in [p_{q-1/2},p_{q+1/2}]$ i.e. $p_{\ell+1/2}-at \in C_q$, then we have

\[f_{j,\ell}(t) =\frac{1}{\Delta p} \int_{p_{q-1/2}-at} ^{p_{q-1/2}} f(x_j,p,0)\mathrm{d}{p}+f_{j,q}(0)-\frac{1}{\Delta p} \int_{p_{q+1/2}} ^{p_{q+1/2}-at} f(x_j,p,0)\mathrm{d}{p}.\]

Here we need to reconstruct a polynomial function $f(x_j,p,0)$ using the averages $f_{j,l}(0)$ using the PSM approach.

VectorSpin.bspline — Method

bspline(p, j, x)

Return the value at x in [0,1] of the B-spline with integer nodes of degree p with support starting at j. Implemented recursively using the De Boor's Algorithm

\[B_{i,0}(x) := \left\{ \begin{matrix} 1 & \mathrm{if} \quad t_i ≤ x < t_{i+1} \\ 0 & \mathrm{otherwise} \end{matrix} \right.\]

\[B_{i,p}(x) := \frac{x - t_i}{t_{i+p} - t_i} B_{i,p-1}(x) + \frac{t_{i+p+1} - x}{t_{i+p+1} - t_{i+1}} B_{i+1,p-1}(x).\]

VectorSpin.init — Method

init(k, x, i, v1int, frequency, a, femi, tiK)

Initialize the distribution function

VectorSpin.integrate — Method

integrate(value, n)

computation of integral average in each cell using newton-cotes formula

VectorSpin.numeint — Method

computation of integral average in each cell using newton-cotes formula

numeint(value, N)

VectorSpin.step! — Method

step!(op, f0, f1, f2, f3, E1, dt)

\[\begin{aligned} \dot{x} & =p \\ \dot{E}_x & = - \int (p f ) dp ds \end{aligned}\]

$H_p$ operator

VectorSpin.step! — Method

compute the subsystem H2

step!(op, f0, f1, f2, f3, E3, A3, dt, h_int)

VectorSpin.step! — Method

step!(op, f0, f1, f2, f3, E2, A2, dt, h_int)

VectorSpin.step! — Method

step!(op, f0, f1, f2, f3, E2, E3, A2, A3, dt)

\[\begin{aligned} \dot{p} = (A_y, A_z) \cdot \partial_x (A_y, A_z) \\ \dot{Ey} = -\partial_x^2 A_y + A_y \rho \\ \dot{Ez} = -\partial_x^2 A_z + A_z \rho \\ \end{aligned}\]

VectorSpin.step! — Method

step!(op, f0, f1, f2, f3, E1, E2, E3, A2, A3, dt)

compute the first subsystem He()

VectorSpin.step! — Method

step!(op, f0, f1, f2, f3, E1, dt)

subsystem for Hv:

\[f_t + vf_x = 0\]

\[-E_x = \rho - 1\]

VectorSpin.translation! — Method

translation!(df, delta, mesh)

interpolate df(x - delta)